requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
pinecone>=3.0.0
streamlit>=1.22.0
python-dotenv>=1.0.0
openai>=1.0.0
tiktoken>=0.5.0
tenacity>=8.0.0
markdown>=3.5.0
pdfkit>=1.0.0
watchdog>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0 